pyproject.toml
[project]
name = "fullstack-project-tracker-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "djangorestframework>=3.17.1",
    "pygments>=2.20.0",
]

[dependency-groups]
dev = [
    "httpie>=3.2.4",
    "ruff>=0.15.12",
]
